update  unit tests
